+++ .circleci/check-markdown-links.sh
@@ -2,6 +2,22 @@ set -euo pipefail
 script_dirpath="$(cd "$(dirname "${0}")" && pwd)"
 root_dirpath="$(dirname "${script_dirpath}")"
 
+# We can use relative syntax to specify kurtosis-core/kurtosis-libs in URLs because everything gets
+# published to the same docs.kurtosistech.com domain, but when checking Markdown links we need to expand
+# the full domain name
+CONFIG_JSON='{
+    "replacementPatterns": [
+        { 
+            "pattern": "^./kurtosis-core",
+            "replacement": "https://docs.kurtosistech.com/kurtosis-core"
+        },
+        { 
+            "pattern": "^./kurtosis-libs", 
+            "replacement": "https://docs.kurtosistech.com/kurtosis-libs"
+        }
+    ]
+}'
+
 # Inspired by https://github.com/open-telemetry/opentelemetry-collector/pull/1156/files/2244e61f4dd0378deffc00d939edf6f800687dcf
 exit_code=0
 for filepath in $(find "${root_dirpath}" -iname '*.md' | sort); do