+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # TBD
 ### Features
 * Added this changelog
+
+### Fixes
+* Relative paths to `kurtosis-core` and `kurtosis-libs` (e.g. `./kurtosis-core`) now get fully expanded when checking Markdown links